Make knowledge source details card usable on touch devices (#24)
Previously, one could not access the knowledge source detail card on
touch devices, because there is no hover and a click opened the source
file.
With this change, the card will open on click on touch devices, and a
"file symlink" button allows to access the file from within the card. A
click outside of the card will close it again.