Implement jump to & highlight source when opening PDF (#29)
Resolves #22.

With this change, the PDF viewer will now jump to and highlight the text
of the source that has been clicked to open the PDF.
This involves several implementation steps:
- Support passing initial page and highlight query to PDF viewer.
- Pass the source page & text as as initial props the PDF viewer when
opening it.
- Ignore line breaks and extra whitespace when highlighting search query
in PDF viewer. Whilst this highlights some additional stuff, it is
needed to highlight the full source text.
- Only use the highlight renderer on the initial page and search result
pages for better performance.
- Normalize text and search query when searching, i.e. remove line
breaks and extra whitespace, to improve search results.
- Improve performance of search and highlighting and cache results and
compiled regexes.
- Improve quality of highlighting.

---------

Co-authored-by: Felix Hertweck <github@felix-hertweck.de>